Moved IndexingService to persistence module
