+++ services/src/main/java/edu/unc/lib/dl/cdr/services/rest/modify/IndexingController.java
@@ -31,10 +31,10 @@ import org.springframework.web.bind.annotation.RequestParam;
 
 import edu.unc.lib.dl.acl.exception.AccessRestrictionException;
 import edu.unc.lib.dl.acl.util.AgentPrincipals;
-import edu.unc.lib.dl.cdr.services.processing.IndexingService;
 import edu.unc.lib.dl.fcrepo4.PIDs;
 import edu.unc.lib.dl.fedora.AuthorizationException;
 import edu.unc.lib.dl.fedora.PID;
+import edu.unc.lib.dl.persist.services.indexing.IndexingService;
 
 /**
  * API controller for reindexing the repository
